refactor: introduce app shell layout
